+++ Color/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog for color
 
-## 0.0.0
+## 0.1.1
+
+* Addition of `Graphics.Color.Illuminant.Wikipedia` module
+* Addition of `convertColor` and `convertColorFloat` functions
+
+## 0.1.0
 
 Initial alpha release